CI updates + attempt to fix oss-fuzz fuzzing
