+++ circle.yml
@@ -19,9 +19,9 @@ commands:
 
 jobs:
 
-  go121:
+  go122:
     docker:
-      - image: cimg/go:1.21
+      - image: cimg/go:1.22
     steps:
       - run:
           name: "Install tools"
@@ -44,8 +44,14 @@ jobs:
       - run:
           name: "Fuzzing"
           command: |
-            GOCACHE=/home/circleci/project/corpus-v3 go test . -run - -fuzz FuzzBase10StringCompare -fuzztime 30s
-            GOCACHE=/home/circleci/project/corpus-v3 go test . -run - -fuzz FuzzDecimal -fuzztime 30s
+            GOCACHE=/home/circleci/project/corpus-v3 go test . -run - -fuzz FuzzBinaryOperations -fuzztime 20s
+            GOCACHE=/home/circleci/project/corpus-v3 go test . -run - -fuzz FuzzCompareOperations -fuzztime 20s
+            GOCACHE=/home/circleci/project/corpus-v3 go test . -run - -fuzz FuzzTernaryOperations -fuzztime 20s
+            GOCACHE=/home/circleci/project/corpus-v3 go test . -run - -fuzz FuzzBase10StringCompare -fuzztime 10s
+            GOCACHE=/home/circleci/project/corpus-v3 go test . -run - -fuzz FuzzDecimal -fuzztime 10s
+            GOCACHE=/home/circleci/project/corpus-v3 go test . -run - -fuzz FuzzFloat64 -fuzztime 10s
+            GOCACHE=/home/circleci/project/corpus-v3 go test . -run - -fuzz FuzzLog10 -fuzztime 10s
+            GOCACHE=/home/circleci/project/corpus-v3 go test . -run - -fuzz FuzzSetString -fuzztime 10s
       - save_cache:
           key: corpus-v3-{{ epoch }}
           paths:
@@ -65,7 +71,7 @@ jobs:
 
   bigendian:
     docker:
-      - image: circleci/buildpack-deps:bullseye
+      - image: cimg/base
     steps:
       - run:
           name: "Install QEMU"
@@ -75,6 +81,13 @@ jobs:
       - run:
           name: "Test (PPC64 emulation)"
           command: qemu-ppc64-static uint256.test.ppc64 -test.v
+
+  go121:
+    docker:
+      - image: cimg/go:1.21
+    steps:
+      - checkout
+      - test
 
   go120:
     docker:
@@ -96,9 +109,10 @@ workflows:
   version: 2
   uint256:
     jobs:
+      - go122
       - go121
       - go120
       - go119
       - bigendian:
           requires:
-            - go121
+            - go122
